Add PremiumClient #1
Implementation of Premium Customer Functionality and Points System

Details:
The objective of this implementation is to introduce premium customer
functionality into the system, enhancing the sales process. Prem	ium
customers are distinguished by their ability to earn points based on
their purchase amount, whereas regular customers do not have access to
this feature. During a sale, the system will prompt whether the customer
is a premium member or not. If the customer is identified as premium,
they will accumulate points based on their spending. Specifically, for
every 10 euros spent, the customer earns one point. This incentivizes
customer loyalty and encourages higher spending among premium members.

Approach:
In broad terms, we have implemented the premium customer functionality
within the "ClientePremium" class. When processing a sale, the
overridden method "pay" is invoked, which takes an "Amount" object
representing the total purchase amount. Within this method, the
following actions are performed:

1. The numeric value of the purchase amount is retrieved from the
"Amount" object and stored in a local variable called "amountToPay".
2. The base class's payment method is called with the "totalAmount"
object as an argument, and the resulting payment status is stored.
3. The number of points earned for the transaction is calculated by
dividing the total transaction value by 10 and converting it to an
integer. These earned points are then added to an instance variable
named "points" (assuming it exists in the current class).
4. The number of points earned for the transaction is printed on the
console.
5. Finally, the method returns the payment status, indicating whether
the payment was successful.